requirements.txt
pydantic
fastapi
pyinstaller
setuptools
uuid
uvicorn
python-multipart
pillow
python-dotenv
requests
pandas
torch
torchvision
numpy
python-dateutil
scipy
simpletransformers
pytz